Remove ContentsApi, move getMultipleContents() to TreeApi + adopt nessie CLI
Related Nessie CLI update to list contents with the associated key for both textual and JSON
representation + indicate keys that were not found (instead of a bare Python exception).

Nit-fixes:
* use default-branch for Nessie CLI contents in all cases
* error out with appropriate message when the expected-hash is not present
* add missing `allow_none=True` in `model.py`
