+++ clients/client/src/main/java/org/projectnessie/client/http/v1api/HttpGetContents.java
@@ -52,7 +52,7 @@ final class HttpGetContents extends BaseHttpOnReferenceRequest<GetContentsBuilde
   @Override
   public Map<ContentsKey, Contents> get() throws NessieNotFoundException {
     MultiGetContentsResponse resp =
-        client.getContentsApi().getMultipleContents(refName, hashOnRef, request.build());
+        client.getTreeApi().getMultipleContents(refName, hashOnRef, request.build());
     return resp.getContents().stream()
         .collect(Collectors.toMap(ContentsWithKey::getKey, ContentsWithKey::getContents));
   }
