+++ python/pynessie/utils/__init__.py
@@ -18,6 +18,7 @@
 """Top-level package for Utility module."""
 
 from .content_util import contents_key
+from .content_util import contents_key_as_string
 from .content_util import format_key
 from .expression_util import build_query_expression_for_commit_log_flags
 from .expression_util import build_query_expression_for_contents_listing_flags
@@ -26,6 +27,7 @@ from .expression_util import parse_to_iso8601
 __all__ = [
     "format_key",
     "contents_key",
+    "contents_key_as_string",
     "build_query_expression_for_commit_log_flags",
     "build_query_expression_for_contents_listing_flags",
     "parse_to_iso8601",
