+++ python/pynessie/utils/content_util.py
@@ -18,6 +18,7 @@
 """Utility functions that can be used to manipulate contents like keys and values."""
 
 import re
+from typing import List
 
 from ..model import ContentsKey
 
@@ -27,6 +28,11 @@ _dot_regex = re.compile('\\.(?=([^"]*"[^"]*")*[^"]*$)')
 def format_key(raw_key: str) -> str:
     """Format given key based on regex."""
     elements = _dot_regex.split(raw_key)
+    return contents_key_as_string(elements)
+
+
+def contents_key_as_string(elements: List[str]) -> str:
+    """Content-key elements to string formatting."""
     return ".".join(i.replace(".", "\0") for i in elements if i)
