+++ servers/jax-rs-testextension/src/main/java/org/projectnessie/jaxrs/ext/NessieJaxRsExtension.java
@@ -37,7 +37,6 @@ import org.projectnessie.services.rest.NessieExceptionMapper;
 import org.projectnessie.services.rest.NessieJaxRsJsonMappingExceptionMapper;
 import org.projectnessie.services.rest.NessieJaxRsJsonParseExceptionMapper;
 import org.projectnessie.services.rest.RestConfigResource;
-import org.projectnessie.services.rest.RestContentsResource;
 import org.projectnessie.services.rest.RestTreeResource;
 import org.projectnessie.services.rest.ValidationExceptionMapper;
 import org.projectnessie.versioned.PersistVersionStoreExtension;
@@ -82,7 +81,6 @@ public class NessieJaxRsExtension implements BeforeAllCallback, AfterAllCallback
             ResourceConfig config = new ResourceConfig();
             config.register(RestConfigResource.class);
             config.register(RestTreeResource.class);
-            config.register(RestContentsResource.class);
             config.register(ConfigApiImpl.class);
             config.register(ContentsKeyParamConverterProvider.class);
             config.register(InstantParamConverterProvider.class);
